fix: :hammer: github handler 403

+++ discovery/handlers/api/github.py
@@ -50,64 +50,65 @@ class GHHandler(APIBaseHandler):
         or get all public repos
         """
         user = json.loads(self.get_secure_cookie("user"))
-        try:
-            g = Github(user['access_token'])
-            auth_user = g.get_user()
-            if not repo_name:
-                repo_list = []
-                if auth_user.login:
-                    try:
-                        repos = auth_user.get_repos()
-                    except GithubException:
-                        self.finish({
-                            'success': True,
-                            'exists': False,
-                            'msg': repo_list
-                        })
-                    except Exception as exc:  # unexpected
-                        raise HTTPError(500, reason=str(exc))
-                    else:
-                        repo_list = [repo.name for repo in repos]
+        # if login method was not GitHub this is not allowed
+        # token will not be available for other login methods
+        if not 'access_token' in user: 
+            raise HTTPError(403, reason="Must login with")
+        g = Github(user['access_token'])
+        auth_user = g.get_user()
+        if not repo_name:
+            repo_list = []
+            if auth_user.login:
+                try:
+                    repos = auth_user.get_repos()
+                except GithubException:
+                    self.finish({
+                        'success': True,
+                        'exists': False,
+                        'msg': repo_list
+                    })
+                except Exception as exc:  # unexpected
+                    raise HTTPError(500, reason=str(exc))
+                else:
+                    repo_list = [repo.name for repo in repos]
+                    self.finish({
+                        'success': True,
+                        'exists': True,
+                        'msg': repo_list
+                    })
+            else:
+                raise HTTPError(400, reason="Unable to authenticate user")
+        else:
+            if auth_user.login:
+                try:
+                    repo = auth_user.get_repo(repo_name)
+                except GithubException.UnknownObjectException:
+                    self.finish({
+                        'success': True,
+                        'exists': False,
+                        'msg': f"{repo_name} does not exist"
+                    })
+                except Exception as exc:  # unexpected
+                    raise HTTPError(500, reason=str(exc))
+                else:
+                    if repo.full_name:
+                        contents = repo.get_contents("")
+                        files = []
+                        for content_file in contents:
+                            files.append(content_file.path)
                         self.finish({
                             'success': True,
                             'exists': True,
-                            'msg': repo_list
+                            'msg': files
                         })
-                else:
-                    raise HTTPError(400, reason="Unable to authenticate user")
-            else:
-                if auth_user.login:
-                    try:
-                        repo = auth_user.get_repo(repo_name)
-                    except GithubException.UnknownObjectException:
+                    else:
                         self.finish({
                             'success': True,
                             'exists': False,
                             'msg': f"{repo_name} does not exist"
                         })
-                    except Exception as exc:  # unexpected
-                        raise HTTPError(500, reason=str(exc))
-                    else:
-                        if repo.full_name:
-                            contents = repo.get_contents("")
-                            files = []
-                            for content_file in contents:
-                                files.append(content_file.path)
-                            self.finish({
-                                'success': True,
-                                'exists': True,
-                                'msg': files
-                            })
-                        else:
-                            self.finish({
-                                'success': True,
-                                'exists': False,
-                                'msg': f"{repo_name} does not exist"
-                            })
-                else:
-                    raise HTTPError(400, reason="Unable to authenticate user")
-        except KeyError:
-            raise HTTPError(403, reason="Must login with GitHub account")
+            else:
+                raise HTTPError(400, reason="Unable to authenticate user")
 
     @authenticated
     def delete(self):
@@ -118,28 +119,29 @@ class GHHandler(APIBaseHandler):
         }
         """
         user = json.loads(self.get_secure_cookie("user"))
+        # if login method was not GitHub this is not allowed
+        # token will not be available for other login methods
+        if not 'access_token' in user: 
+            raise HTTPError(403, reason="Must login with")
+        g = Github(user['access_token'])
+        # authenticated user
+        auth_user = g.get_user()
+        repo_name = self.args_json.get('name', '')
+        if not repo_name:
+            raise HTTPError(400, reason="Repo name not provided")
         try:
-            g = Github(user['access_token'])
-            # authenticated user
-            auth_user = g.get_user()
-            repo_name = self.args_json.get('name', '')
-            if not repo_name:
-                raise HTTPError(400, reason="Repo name not provided")
-            try:
-                repo = auth_user.get_repo(full_name=repo_name)
-                repo.delete()
-            except GithubException as e:
-                raise HTTPError(400, reason=e)
-            except Exception as exc:  # unexpected
-                raise HTTPError(500, reason=str(exc))
-            if not repo:
-                logging.info("Deleted repo %s", repo_name)
-                self.finish({
-                    'success': True,
-                    'msg': f"{repo_name} was deleted"
-                })
-        except KeyError:
-            raise HTTPError(403, reason="Must login with GitHub account")
+            repo = auth_user.get_repo(full_name=repo_name)
+            repo.delete()
+        except GithubException as e:
+            raise HTTPError(400, reason=e)
+        except Exception as exc:  # unexpected
+            raise HTTPError(500, reason=str(exc))
+        if not repo:
+            logging.info("Deleted repo %s", repo_name)
+            self.finish({
+                'success': True,
+                'msg': f"{repo_name} was deleted"
+            })
 
     @authenticated
     def post(self):
@@ -167,50 +169,50 @@ class GHHandler(APIBaseHandler):
 
         logging.info("Repo name %s", repo_name)
         user = json.loads(self.get_secure_cookie("user"))
-        # signin with token
-        try:
-            g = Github(user['access_token'])
-            # authenticated user
-            auth_user = g.get_user()
-            if auth_user.login:
+        # if login method was not GitHub this is not allowed
+        # token will not be available for other login methods
+        if not 'access_token' in user: 
+            raise HTTPError(403, reason="Must login with")
+        g = Github(user['access_token'])
+        # authenticated user
+        auth_user = g.get_user()
+        if auth_user.login:
 
-                if existing:
-                    try:
-                        repo = auth_user.get_repo(repo_name)
-                    except GithubException.UnknownObjectException:
-                        self.finish({
-                            'success': False,
-                            'exists': True,
-                            'msg': f"Repo {repo_name} already exists"
-                        })
-                    except Exception as exc:  # unexpected
-                        raise HTTPError(500, reason=str(exc))
-                else:
-                    try:
-                        repo = auth_user.create_repo(repo_name)
-                    except GithubException as e:
-                        raise HTTPError(400, reason=e)
-                    except Exception as exc:  # unexpected
-                        raise HTTPError(500, reason=str(exc))
-
-            else:
-                raise HTTPError(400, reason="Unable to authenticate user")
-
-            if repo:
-                path = file_name
-                file = repo.create_file(path, msg, content_encoded)
-                if file:
+            if existing:
+                try:
+                    repo = auth_user.get_repo(repo_name)
+                except GithubException.UnknownObjectException:
                     self.finish({
-                        'success': True,
-                        'msg': repo.full_name
+                        'success': False,
+                        'exists': True,
+                        'msg': f"Repo {repo_name} already exists"
                     })
+                except Exception as exc:  # unexpected
+                    raise HTTPError(500, reason=str(exc))
             else:
+                try:
+                    repo = auth_user.create_repo(repo_name)
+                except GithubException as e:
+                    raise HTTPError(400, reason=e)
+                except Exception as exc:  # unexpected
+                    raise HTTPError(500, reason=str(exc))
+
+        else:
+            raise HTTPError(400, reason="Unable to authenticate user")
+
+        if repo:
+            path = file_name
+            file = repo.create_file(path, msg, content_encoded)
+            if file:
                 self.finish({
-                    'success': False,
-                    'msg': f"{repo_name} does not exist"
+                    'success': True,
+                    'msg': repo.full_name
                 })
-        except KeyError:
-            raise HTTPError(403, reason="Must login with GitHub account")
+        else:
+            self.finish({
+                'success': False,
+                'msg': f"{repo_name} does not exist"
+            })
 
     @authenticated
     def put(self):
@@ -241,44 +243,45 @@ class GHHandler(APIBaseHandler):
 
         logging.info("Updating file on repo.name: %s", repo_name)
         user = json.loads(self.get_secure_cookie("user"))
-        try:
-            g = Github(user['access_token'])
-            # authenticated user
-            auth_user = g.get_user()
-            if auth_user.login:
-                try:
-                    repo = auth_user.get_repo(repo_name)
-                except GithubException as e:
-                    raise HTTPError(400, reason=e)
-                except Exception as exc:  # unexpected
-                    raise HTTPError(500, reason=str(exc))
-            else:
-                raise HTTPError(400, reason="Unable to authenticate user")
-            if repo and existing_file:
-                file_path = file_name
-                try:
-                    file = repo.get_contents(file_path)
-                except GithubException as e:
-                    raise HTTPError(400, reason=e)
-                except Exception as exc:  # unexpected
-                    raise HTTPError(500, reason=str(exc))
-                if file:
-                    file_updated = repo.update_file(file_path, msg, content_encoded, file.sha)
-                    if file_updated:
-                        logging.info("File updated name %s", file_name)
-                        self.finish({
-                            'success': True,
-                            'msg': repo.full_name
-                        })
-                    else:
-                        self.finish({
-                            'success': False,
-                            'msg': repo.full_name
-                        })
-            else:
-                self.finish({
-                    'success': False,
-                    'msg': f"{repo_name} repo or file must exist."
-                })
-        except KeyError:
-            raise HTTPError(403, reason="Must login with GitHub account")
+        # if login method was not GitHub this is not allowed
+        # token will not be available for other login methods
+        if not 'access_token' in user: 
+            raise HTTPError(403, reason="Must login with")
+        g = Github(user['access_token'])
+        # authenticated user
+        auth_user = g.get_user()
+        if auth_user.login:
+            try:
+                repo = auth_user.get_repo(repo_name)
+            except GithubException as e:
+                raise HTTPError(400, reason=e)
+            except Exception as exc:  # unexpected
+                raise HTTPError(500, reason=str(exc))
+        else:
+            raise HTTPError(400, reason="Unable to authenticate user")
+        if repo and existing_file:
+            file_path = file_name
+            try:
+                file = repo.get_contents(file_path)
+            except GithubException as e:
+                raise HTTPError(400, reason=e)
+            except Exception as exc:  # unexpected
+                raise HTTPError(500, reason=str(exc))
+            if file:
+                file_updated = repo.update_file(file_path, msg, content_encoded, file.sha)
+                if file_updated:
+                    logging.info("File updated name %s", file_name)
+                    self.finish({
+                        'success': True,
+                        'msg': repo.full_name
+                    })
+                else:
+                    self.finish({
+                        'success': False,
+                        'msg': repo.full_name
+                    })
+        else:
+            self.finish({
+                'success': False,
+                'msg': f"{repo_name} repo or file must exist."
+            })
